create position for page Statistic
